🎨 Improve Movie related components and folder structure
