+++ apps/web/src/components/movie/MovieActionsModal/components/AddMovieToListModal/components/CreateMovieListModal/types.ts
@@ -1,0 +1,26 @@
+import type { SchemaOf } from 'yup';
+
+import type { ModalHandles } from '../../../../../../Modal';
+
+import type { CreateUserListMutationResult } from '../../../../../../../graphql';
+
+interface CreateMovieListFormValues {
+  listName: string;
+  description?: string;
+}
+
+export type ValidationSchemaType = SchemaOf<CreateMovieListFormValues>;
+
+export type SubmitHandles = (
+  values: CreateMovieListFormValues,
+) => Promise<void>;
+
+export type CreateMovieListModalLogicProps = ModalHandles;
+
+export interface CreateMovieListModalLogicHandles {
+  createUserListResult: CreateUserListMutationResult;
+
+  validationSchema: ValidationSchemaType;
+
+  handleSubmit: SubmitHandles;
+}
