+++ apps/web/src/views/movies/root/index.tsx
@@ -1,12 +1,8 @@
-import { useMemo } from 'react';
-
 import { useRouter } from 'next/router';
 
-import type { Movie } from '../../../graphql';
+import { Movie } from '../../../graphql';
 
-import type { MovieContextData } from './hooks/useMovie';
-
-import { MovieContext } from './hooks/useMovie';
+import { MovieContextProvider } from './hooks/useMovie';
 
 import { useAuth } from '../../../hooks/useAuth';
 
@@ -14,97 +10,82 @@ import Card from '../../../components/Card';
 
 import PageContent from '../../../components/PageContent';
 
-import BackgroundImage from '../../../components/BackgroundImage';
-
 import TextShorter from '../../../components/TextShorter';
-
-import Link from '../../../components/Link';
 
 import OptionsButton from './components/OptionsButton';
 
 import MovieCast from './components/MovieCast';
 
-import MovieViewSkeleton from '../view/Skeleton';
-
-import MovieHeader from '../view/Header';
+import Button from '../../../components/Button';
 
 import RecentReviews from './components/Reviews/RecentReviews';
 
-import PopularReviews from './components/Reviews/PopularReviews';
 import BackdropImage from '../../../components/BackdropImage';
+
+import MovieInfos from '../../../components/movie/MovieInfos';
+
+import PopularReviews from './components/Reviews/PopularReviews';
+
+import MovieGenres from './components/MovieGenres';
+
+import MovieInfosSkeleton from '../../../components/movie/MovieInfos/Skeleton';
 
 interface MovieViewProps {
   movie: Movie;
 }
 
 const MovieView: React.FC<MovieViewProps> = ({ movie }) => {
-  const { user } = useAuth();
-
-  const contextProviderValue = useMemo(
-    () => ({ movie } as MovieContextData),
-    [movie],
-  );
+  const { data: session } = useAuth();
 
   const { isFallback } = useRouter();
 
   if (isFallback) {
-    return <MovieViewSkeleton />;
-  }
-
-  if (!movie) {
-    return <h1 className="text-danger-base">Movie not found</h1>;
+    return <MovieInfosSkeleton />;
   }
 
   return (
-    <BackdropImage src={movie.backdropUrl} alt="Movie backdrop image">
+    <BackdropImage src={movie.backdropUrl} alt={movie.originalTitle}>
       <PageContent>
-        <MovieContext.Provider value={contextProviderValue}>
-          <MovieHeader movie={movie} />
+        <MovieContextProvider movie={movie}>
+          <MovieInfos movie={movie}>
+            <div className="flex items-center gap-1 text-grey-200 mt-2">
+              <Button
+                className="text-sm"
+                buttonStyle="secondary"
+                buttonSize="xs"
+                full={false}
+              >
+                Watch trailer
+              </Button>
 
-          <div className="flex flex-col gap-4">
+              <div className="rounded-md border border-grey-700 px-1 bg-opacity-60">
+                <span>18</span>
+              </div>
+            </div>
+          </MovieInfos>
+
+          <article className="flex flex-col gap-4 mt-4">
             <TextShorter
-              className="text-grey-200 mt-4"
+              className="text-grey-200"
               text={movie.overview}
               maxCharacters={200}
             />
 
-            <div className="flex gap-2 flex-wrap">
-              {movie.genres.map(genre => (
-                <Link
-                  className="py-1 px-2 border rounded-md border-grey-700 transition-colors hover:bg-grey-600"
-                  key={`genre-${genre.id}`}
-                  href="/"
-                >
-                  <h1 className="text-grey-100 whitespace-nowrap">
-                    {genre.name}
-                  </h1>
-                </Link>
-              ))}
-            </div>
+            <MovieGenres />
 
-            {user && <OptionsButton />}
+            {session && <OptionsButton />}
 
             <MovieCast />
 
-            <PopularReviews movieId={movie.id} />
+            <PopularReviews />
 
             <RecentReviews />
 
-            <Card title="Similar movies" noPadding>
-              {/* <div className="flex gap-2">
-    {movieList.map(m => (
-      <Link key={m.id} className="hover:opacity-60" href="/">
-        <MovieCover coverUrl={m.coverUrl} />
-      </Link>
-    ))}
-  </div> */}
-            </Card>
+            <Card title="Similar movies" noPadding />
 
-            <Card title="Popular lists" noPadding>
-              {/* <UserMovieList /> */}
-            </Card>
-          </div>
-        </MovieContext.Provider>
+            <Card title="Popular lists" noPadding />
+          </article>
+        </MovieContextProvider>
       </PageContent>
     </BackdropImage>
   );
