--- apps/web/src/views/movies/view/Header/index.tsx
@@ -1,101 +1,0 @@
-import { parseISO, intervalToDuration } from 'date-fns';
-
-import type { PropsWithChildren } from 'react';
-
-import type { Movie } from '../../../../graphql';
-
-import MovieCover from '../../components/Cover';
-
-import Button from '../../../../components/Button';
-
-import MovieRatingStar from '../../components/RatingStar';
-
-interface MovieHeaderProps {
-  movie: Movie;
-  basicInfo?: boolean;
-}
-
-const MovieHeader: React.FC<PropsWithChildren<MovieHeaderProps>> = ({
-  movie,
-  basicInfo,
-  children,
-}) => {
-  function toHoursAndMinutes(totalMinutes: number) {
-    const minutes = totalMinutes % 60;
-
-    const hours = Math.floor(totalMinutes / 60);
-
-    if (minutes === 0) {
-      return `${hours}h`;
-    }
-
-    return `${hours}h ${minutes}m`;
-  }
-
-  const movieDate = parseISO(movie.releaseDate);
-
-  // TODO change that, not showing #1 director
-  const directors = movie.credits.crew.filter(
-    crew => crew.department === 'Directing',
-  );
-
-  return (
-    <div className="flex justify-between gap-2">
-      <div className="flex flex-col w-full z-10">
-        <h1 className="text-grey-100 text-2xl font-semibold">
-          {movie.originalTitle}
-        </h1>
-
-        <div className="text-grey-200 mb-2">
-          <div className="flex items-center gap-1">
-            {!basicInfo && (
-              <>
-                <MovieRatingStar color="yellow" rating={8.3} checked />
-
-                <span>•</span>
-              </>
-            )}
-
-            <span className="text-sm">
-              {movieDate.getMonth()}/{movieDate.getDay()}/
-              {movieDate.getFullYear()}
-            </span>
-
-            <span>•</span>
-
-            <span className="text-sm">{toHoursAndMinutes(movie.runtime)}</span>
-          </div>
-
-          {directors.length > 0 && (
-            <span className="text-sm mr-1">
-              Directed by <strong>{directors[0].originalName}</strong>
-            </span>
-          )}
-        </div>
-
-        {!basicInfo && (
-          <div className="flex items-center gap-1 text-grey-200 mt-auto">
-            <Button
-              className="text-sm"
-              buttonStyle="secondary"
-              buttonSize="xs"
-              full={false}
-            >
-              Watch trailer
-            </Button>
-
-            <div className="rounded-md border border-grey-700 px-1 bg-opacity-60">
-              <span>18</span>
-            </div>
-          </div>
-        )}
-
-        {children}
-      </div>
-
-      <MovieCover coverUrl={movie.posterUrl} />
-    </div>
-  );
-};
-
-export default MovieHeader;
