--- apps/web/src/views/movies/view/Skeleton/index.tsx
@@ -1,11 +1,0 @@
-import MovieBodySkeleton from '../Body/Skeleton';
-
-import MovieHeaderSkeleton from '../Header/Skeleton';
-
-const MovieViewSkeleton: React.FC = () => (
-  <MovieHeaderSkeleton>
-    <MovieBodySkeleton />
-  </MovieHeaderSkeleton>
-);
-
-export default MovieViewSkeleton;
